Fix bug when reading mass that have block
A bug was introduced that only read the mass from the mass table in the
header, which resulted in arrays of only zeros for particle types that
actually have real mass blocks instead of an entry in the header mass
table.
Starting with this change it also throws an error when the respective
block is present, but not for the specified particle type.