Fix bug for parttype=-1; remove mass special case
When reading particles with parttype=-1, the properties were also read
for particles that don't even have that property, leading to wrong
values being returned.

Due to complications, the special casing for the MASS block has now been
removed. An error is thrown when the MASS block is missing, even if
there is an entry in the massarr property of the header. The user has to
perform that check in their own code now.